requirements.txt
streamlit
PyPDF2
weaviate-client
openai
python-dotenv